+++ src/main/java/de/maxhenkel/shulkerbox/Main.java
@@ -1,8 +1,9 @@
 package de.maxhenkel.shulkerbox;
 
+import de.maxhenkel.corelib.ClientRegistry;
+import de.maxhenkel.corelib.CommonRegistry;
 import de.maxhenkel.shulkerbox.gui.ShulkerboxContainer;
 import de.maxhenkel.shulkerbox.gui.ShulkerboxScreen;
-import net.minecraft.client.gui.ScreenManager;
 import net.minecraft.inventory.container.ContainerType;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.api.distmarker.Dist;
@@ -11,7 +12,6 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.DistExecutor;
-import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
@@ -24,13 +24,13 @@ public class Main {
     public static final String MODID = "shulkerbox";
 
     public static ContainerType SHULKERBOX_CONTAINER;
+    public static ServerConfig SERVER_CONFIG;
 
     public Main() {
         FMLJavaModLoadingContext.get().getModEventBus().addListener(this::commonSetup);
         FMLJavaModLoadingContext.get().getModEventBus().addGenericListener(ContainerType.class, this::registerContainers);
 
-        ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, Config.SERVER_SPEC);
-        ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, Config.CLIENT_SPEC);
+        SERVER_CONFIG = CommonRegistry.registerConfig(ModConfig.Type.SERVER, ServerConfig.class);
 
         DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> FMLJavaModLoadingContext.get().getModEventBus().addListener(Main.this::clientSetup));
     }
@@ -44,8 +44,7 @@ public class Main {
     @SubscribeEvent
     @OnlyIn(Dist.CLIENT)
     public void clientSetup(FMLClientSetupEvent event) {
-        ScreenManager.IScreenFactory factory = (ScreenManager.IScreenFactory<ShulkerboxContainer, ShulkerboxScreen>) (container, playerInventory, name) -> new ShulkerboxScreen(playerInventory, container, name);
-        ScreenManager.registerFactory(Main.SHULKERBOX_CONTAINER, factory);
+        ClientRegistry.<ShulkerboxContainer, ShulkerboxScreen>registerScreen(Main.SHULKERBOX_CONTAINER, ShulkerboxScreen::new);
     }
 
     @SubscribeEvent
@@ -54,4 +53,5 @@ public class Main {
         SHULKERBOX_CONTAINER.setRegistryName(new ResourceLocation(Main.MODID, "shulkerbox"));
         event.getRegistry().register(SHULKERBOX_CONTAINER);
     }
+
 }