+++ src/main/java/de/maxhenkel/shulkerbox/gui/ShulkerboxContainer.java
@@ -1,20 +1,22 @@
 package de.maxhenkel.shulkerbox.gui;
 
+import de.maxhenkel.corelib.inventory.ContainerBase;
+import de.maxhenkel.corelib.inventory.LockedSlot;
+import de.maxhenkel.shulkerbox.AdvancedShulkerBoxInventory;
 import de.maxhenkel.shulkerbox.Main;
-import de.maxhenkel.shulkerbox.ShulkerBoxInventory;
 import de.maxhenkel.shulkerbox.Utils;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.IInventory;
 import net.minecraft.inventory.container.ShulkerBoxSlot;
+import net.minecraft.inventory.container.Slot;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.Hand;
 
 public class ShulkerboxContainer extends ContainerBase {
 
-    private IInventory albumInventory;
-
-    public ShulkerboxContainer(int id, IInventory playerInventory, IInventory shulkerboxInventory) {
+    public ShulkerboxContainer(int id, PlayerInventory playerInventory, IInventory shulkerboxInventory) {
         super(Main.SHULKERBOX_CONTAINER, id, playerInventory, shulkerboxInventory);
-        this.albumInventory = shulkerboxInventory;
 
         for (int y = 0; y < 3; y++) {
             for (int x = 0; x < 9; x++) {
@@ -22,25 +24,34 @@ public class ShulkerboxContainer extends ContainerBase {
             }
         }
 
-        addInvSlots();
+        for (int l = 0; l < 3; l++) {
+            for (int j1 = 0; j1 < 9; j1++) {
+                addSlot(new Slot(playerInventory, j1 + l * 9 + 9, 8 + j1 * 18, 102 + l * 18 - 18));
+            }
+        }
+
+        int locked = getLockedSlot(playerInventory.player);
+        for (int i = 0; i < 9; i++) {
+            int x = 8 + i * 18;
+            int y = 142;
+            if (i == locked) {
+                addSlot(new LockedSlot(playerInventory, i, x, y));
+            } else {
+                addSlot(new Slot(playerInventory, i, x, y));
+            }
+        }
     }
 
     public ShulkerboxContainer(int id, PlayerInventory playerInventory) {
-        this(id, playerInventory, new ShulkerBoxInventory(playerInventory.player, Utils.getShulkerBox(playerInventory.player)));
+        this(id, playerInventory, new AdvancedShulkerBoxInventory(playerInventory.player, Utils.getShulkerBox(playerInventory.player)));
     }
 
-    @Override
-    public int getInventorySize() {
-        return 27;
+    public static int getLockedSlot(PlayerEntity player) {
+        ItemStack stack = player.getHeldItem(Hand.MAIN_HAND);
+        if (Utils.isShulkerBox(stack)) {
+            return player.inventory.currentItem;
+        }
+        return -1;
     }
 
-    @Override
-    public int getInvOffset() {
-        return 0;
-    }
-
-    @Override
-    public boolean canInteractWith(PlayerEntity playerIn) {
-        return albumInventory.isUsableByPlayer(playerIn);
-    }
 }