+++ src/main/java/de/maxhenkel/shulkerbox/gui/ShulkerboxScreen.java
@@ -1,6 +1,7 @@
 package de.maxhenkel.shulkerbox.gui;
 
 import com.mojang.blaze3d.matrix.MatrixStack;
+import de.maxhenkel.corelib.inventory.ScreenBase;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
@@ -11,7 +12,7 @@ public class ShulkerboxScreen extends ScreenBase<ShulkerboxContainer> {
 
     private PlayerInventory playerInventory;
 
-    public ShulkerboxScreen(PlayerInventory playerInventory, ShulkerboxContainer shulkerboxContainer, ITextComponent name) {
+    public ShulkerboxScreen(ShulkerboxContainer shulkerboxContainer, PlayerInventory playerInventory, ITextComponent name) {
         super(DEFAULT_IMAGE, shulkerboxContainer, playerInventory, name);
 
         this.playerInventory = playerInventory;
@@ -24,4 +25,5 @@ public class ShulkerboxScreen extends ScreenBase<ShulkerboxContainer> {
         field_230712_o_.func_238421_b_(matrixStack, func_231171_q_().getString(), 8F, 6F, FONT_COLOR);
         field_230712_o_.func_238421_b_(matrixStack, playerInventory.getDisplayName().getString(), 8F, (float) (ySize - 96 + 4), FONT_COLOR);
     }
+
 }